pyproject.toml
[project]
name = "commodity-curve-factors"
version = "0.1.0"
description = "Commodity Futures Curve Modeling and Factor Trading"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "yfinance>=0.2.31",
    "fredapi>=0.5",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.7",
    "pandas-stubs>=2.0",
    "types-PyYAML>=6.0",
    "types-requests>=2.31",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.coverage.run]
source = ["src/commodity_curve_factors"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "if __name__ == .__main__.",
    "pragma: no cover",
]
